chore: split up types validation to corresponding type files
