+++ internal/types/instance.go
@@ -2,6 +2,9 @@
 package types
 
 import (
+	"fmt"
+	"strings"
+
 	"github.com/celestiaorg/talis/internal/db/models"
 )
 
@@ -109,4 +112,109 @@ type InstanceInfo struct {
 	Size          string            // Instance size/type
 	Volumes       []string          `json:"volumes,omitempty"`        // List of attached volume IDs
 	VolumeDetails []VolumeDetails   `json:"volume_details,omitempty"` // Detailed information about attached volumes
+}
+
+// Validate validates the infrastructure request
+func (r *InstancesRequest) Validate() error {
+	if r.JobName == "" {
+		return fmt.Errorf("job_name is required")
+	}
+
+	if r.ProjectName == "" {
+		return fmt.Errorf("project_name is required")
+	}
+	if len(r.Instances) == 0 {
+		return fmt.Errorf("at least one instance configuration is required")
+	}
+
+	for i, instance := range r.Instances {
+		if instance.Name == "" && r.InstanceName == "" {
+			return fmt.Errorf("instance_name or instance.name is required")
+		}
+
+		// Validate hostname format
+		nameToValidate := instance.Name
+		if nameToValidate == "" {
+			nameToValidate = r.InstanceName
+		}
+		if err := validateHostname(nameToValidate); err != nil {
+			return fmt.Errorf("invalid hostname at index %d: %w", i, err)
+		}
+
+		if err := instance.Validate(); err != nil {
+			return fmt.Errorf("invalid instance configuration at index %d: %w", i, err)
+		}
+	}
+
+	return nil
+}
+
+// Validate validates the instance configuration
+func (i *InstanceRequest) Validate() error {
+	if i.Provider == "" {
+		return fmt.Errorf("provider is required")
+	}
+	if i.NumberOfInstances < 1 {
+		return fmt.Errorf("number_of_instances must be greater than 0")
+	}
+	if i.Region == "" {
+		return fmt.Errorf("region is required")
+	}
+	if i.Size == "" {
+		return fmt.Errorf("size is required")
+	}
+	if i.Image == "" {
+		return fmt.Errorf("image is required")
+	}
+	if i.SSHKeyName == "" {
+		return fmt.Errorf("ssh_key_name is required")
+	}
+
+	// Validate volumes if present
+	for j, vol := range i.Volumes {
+		if err := ValidateVolume(&vol, i.Region); err != nil {
+			return fmt.Errorf("invalid volume configuration at index %d: %w", j, err)
+		}
+	}
+
+	i.SSHKeyName = strings.ToLower(i.SSHKeyName)
+	return nil
+}
+
+// Validate validates the delete request
+func (r *DeleteRequest) Validate() error {
+	if r.InstanceName == "" {
+		return fmt.Errorf("instance_name is required")
+	}
+	if r.ProjectName == "" {
+		return fmt.Errorf("project_name is required")
+	}
+	if len(r.Instances) == 0 {
+		return fmt.Errorf("at least one instance configuration is required")
+	}
+
+	for i, instance := range r.Instances {
+		if err := instance.Validate(); err != nil {
+			return fmt.Errorf("invalid instance configuration at index %d: %w", i, err)
+		}
+	}
+
+	return nil
+}
+
+// Validate validates the delete instance configuration
+func (i *DeleteInstance) Validate() error {
+	if i.Provider == "" {
+		return fmt.Errorf("provider is required")
+	}
+	if i.NumberOfInstances < 1 {
+		return fmt.Errorf("number_of_instances must be greater than 0")
+	}
+	if i.Region == "" {
+		return fmt.Errorf("region is required")
+	}
+	if i.Size == "" {
+		return fmt.Errorf("size is required")
+	}
+	return nil
 }
